+++ app/services/doc_auth/response.rb
@@ -34,20 +34,20 @@ module DocAuth
       @selfie_quality_good = selfie_quality_good
     end
 
-    def merge(other)
-      Response.new(
-        success: success? && other.success?,
-        errors: errors.merge(other.errors),
-        exception: exception || other.exception,
-        extra: extra.merge(other.extra),
-        pii_from_doc: pii_from_doc.merge(other.pii_from_doc),
-        attention_with_barcode: attention_with_barcode? || other.attention_with_barcode?,
-        doc_type_supported: doc_type_supported? || other.doc_type_supported?,
-        selfie_status: selfie_status,
-        selfie_live: selfie_live?,
-        selfie_quality_good: selfie_quality_good?,
-      )
-    end
+    # def merge(other)
+    #   Response.new(
+    #     success: success? && other.success?,
+    #     errors: errors.merge(other.errors),
+    #     exception: exception || other.exception,
+    #     extra: extra.merge(other.extra),
+    #     pii_from_doc: pii_from_doc.merge(other.pii_from_doc),
+    #     attention_with_barcode: attention_with_barcode? || other.attention_with_barcode?,
+    #     doc_type_supported: doc_type_supported? || other.doc_type_supported?,
+    #     selfie_status: selfie_status,
+    #     selfie_live: selfie_live?,
+    #     selfie_quality_good: selfie_quality_good?,
+    #   )
+    # end
 
     def success?
       @success
