+++ spec/services/doc_auth/lexis_nexis/result_codes_spec.rb
@@ -1,6 +1,6 @@
 require 'rails_helper'
 
-RSpec.describe DocAuth::ResultCodes do
+RSpec.describe DocAuth::LexisNexis::ResultCodes do
   describe '.from_int' do
     it 'is a result code for the int' do
       result_code = DocAuth::ResultCodes.from_int(1)
